Complete Course :)

+++ app/Http/Controllers/Api/ArticleController.php
@@ -6,12 +6,20 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\SaveArticleRequest;
 use App\Http\Resources\ArticleResource;
 use App\Models\Article;
+use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Http\Response;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class ArticleController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum', [
+            'only' => ['store', 'update', 'destroy']
+        ]);
+    }
     public function index(): AnonymousResourceCollection
     {
         $articles = Article::query()
@@ -34,18 +42,24 @@ class ArticleController extends Controller
     }
     public function store(SaveArticleRequest $request): ArticleResource
     {
+        $this->authorize('create', new Article);
+
         $article = Article::create($request->validated());
 
         return ArticleResource::make($article);
     }
     public function update(Article $article, SaveArticleRequest $request): ArticleResource
     {
+        $this->authorize('update', $article);
+
         $article->update($request->validated());
 
         return ArticleResource::make($article);
     }
-    public function destroy(Article $article): Response
+    public function destroy(Article $article, Request $request): Response
     {
+        $this->authorize('delete', $article);
+
         $article->delete();
 
         return response()->noContent();
